ALHAMDULILLAH

+++ routes/web.php
@@ -37,62 +37,59 @@ use SebastianBergmann\CodeCoverage\Report\Html\Dashboard;
 Route::get('/', function () {
     return redirect('/login');
 });
-
-// Route::get('/dashboardapar', [AparInspeksiController::class, 'index']);
-// Route::get('/aparcreate', function () {
-//     return view('apar.create');
-// });
+Route::get('/getKota/{provinsi}', [KelolaParameterController::class, 'getKota']);
 Route::get('/pilih', [DashboardController::class, 'pilih']);
-// Route::group([
-//     'prefix' => config('admin.prefix'),
-//     'namespace' => 'App\\Http\\Controllers',
-// ], function () {
+
 
 Route::get('/login', [LoginAdminController::class, 'formLogin'])->middleware('guest')->name('login');
 Route::post('/login', [LoginAdminController::class, 'login'])->middleware('guest');
 Route::post('/logout', [LoginAdminController::class, 'logout'])->middleware('auth');
-
+Route::get('/editAkun', [UserController::class, 'editAkun'])->middleware('auth');
+Route::post('/editAkun/{id}', [UserController::class, 'updateAkun']);
 //dashboard petugas
 Route::get('/dashboard', [DashboardController::class, 'index']);
 
 //tambah Supervisior    
-Route::post('/user', [UserController::class, 'store']);
-Route::put('/user/update/{id}', [UserController::class, 'update']);
-Route::put('/user/resetPassword/{id}', [UserController::class, 'resetPassword']);
-Route::delete('/user/{id}', [UserController::class, 'destroy']);
+Route::post('/user', [UserController::class, 'store'])->middleware('auth');
+Route::put('/user/update/{id}', [UserController::class, 'update'])->middleware('auth');
+Route::put('/user/resetPassword/{id}', [UserController::class, 'resetPassword'])->middleware('auth');
+Route::delete('/user/{id}', [UserController::class, 'destroy'])->middleware('auth');
+
 // Route Apar
-Route::group(['prefix' => 'apar', 'middleware' => 'auth'], function () {
-    Route::get('/dashboard', [DashboardController::class, 'apar']);
-    Route::get('/dataapar/export', [DataAparController::class, 'export_pdf']);
-    Route::get('/inspeksiTahunan/{id}/export', [AparInspeksiController::class, 'exportTahunan_pdf']);
-    Route::resource('/dataapar', DataAparController::class);
-    Route::get('/masterinspeksi/{id}/export', [MasterInspeksiController::class, 'export_pdf']);
-    Route::resource('/masterinspeksi', MasterInspeksiController::class);
+Route::group(['prefix' => 'apar'], function () {
+    Route::get('/dashboard', [DashboardController::class, 'apar'])->middleware('supervisior');
+    Route::get('/dataapar/export', [DataAparController::class, 'export_pdf'])->middleware('supervisior');
+    Route::get('/inspeksiTahunan/{id}/export', [AparInspeksiController::class, 'exportTahunan_pdf'])->middleware('supervisior');
+    Route::resource('/dataapar', DataAparController::class)->middleware('supervisior');
+    Route::get('/masterinspeksi/{id}/export', [MasterInspeksiController::class, 'export_pdf'])->middleware('supervisior');
+    Route::resource('/masterinspeksi', MasterInspeksiController::class)->middleware('supervisior');
     Route::get('/inspeksi', [AparInspeksiController::class, 'index']);
     Route::get('/inspeksi/{periode}', [AparInspeksiController::class, 'detailInspeksi']);
     Route::get('/inspeksi/{periode}/inputInpeksiApar', [AparInspeksiController::class, 'create']);
-    // Route::get('/statusapar', [AparInspeksiController::class, 'status']);
     Route::get('/inputInpeksiApar', [AparInspeksiController::class, 'create']);
     Route::post('/inputInpeksiApar', [AparInspeksiController::class, 'store']);
-    Route::get('/kelolaParameter', [KelolaParameterController::class, 'index']);
-    Route::post('/tambahTipe', [KelolaParameterController::class, 'tambahTipe']);
-    Route::post('/tambahJenis', [KelolaParameterController::class, 'tambahJenis']);
-    Route::delete('/deleteTipe/{id}', [KelolaParameterController::class, 'deleteTipe']);
-    Route::delete('/deleteJenis/{id}', [KelolaParameterController::class, 'deleteJenis']);
-    Route::put('/editTipe/{id}', [KelolaParameterController::class, 'editTipe']);
-    Route::put('/editJenis/{id}', [KelolaParameterController::class, 'editJenis']);
+    Route::get('/kelolaParameter', [KelolaParameterController::class, 'index'])->middleware('supervisior');
+    Route::post('/tambahTipe', [KelolaParameterController::class, 'tambahTipe'])->middleware('supervisior');
+    Route::post('/tambahJenis', [KelolaParameterController::class, 'tambahJenis'])->middleware('supervisior');
+    Route::post('/tambahZona', [KelolaParameterController::class, 'tambahZona'])->middleware('supervisior');
+    Route::delete('/deleteTipe/{id}', [KelolaParameterController::class, 'deleteTipe'])->middleware('supervisior');
+    Route::delete('/deleteZona/{id}', [KelolaParameterController::class, 'deleteZona'])->middleware('supervisior');
+    Route::delete('/deleteJenis/{id}', [KelolaParameterController::class, 'deleteJenis'])->middleware('supervisior');
+    Route::put('/editTipe/{id}', [KelolaParameterController::class, 'editTipe'])->middleware('supervisior');
+    Route::put('/editJenis/{id}', [KelolaParameterController::class, 'editJenis'])->middleware('supervisior');
+    Route::put('/editZona/{id}', [KelolaParameterController::class, 'editZona'])->middleware('supervisior');
     Route::get('/hasilInspeksi/{id}', [AparInspeksiController::class, 'hasil']);
 
     // Route::get('/report', [ReportAparController::class, 'index']);
 });
 
 Route::group(['prefix' => 'p3k', 'middleware' => 'auth'], function () {
-    Route::get('/dashboard', [DashboardController::class, 'p3k']);
-    Route::get('/datap3k/export', [DataP3KController::class, 'export_pdf']);
-    Route::get('/inspeksiTahunan/{id}/export', [InspeksiP3KController::class, 'exportTahunan_pdf']);
-    Route::resource('/datap3k', DataP3KController::class);
-    Route::get('/masterinspeksi/{id}/export', [MasterInspeksiP3KController::class, 'export_pdf']);
-    Route::resource('/masterinspeksi', MasterInspeksiP3KController::class);
+    Route::get('/dashboard', [DashboardController::class, 'p3k'])->middleware('supervisior');
+    Route::get('/datap3k/export', [DataP3KController::class, 'export_pdf'])->middleware('supervisior');
+    Route::get('/inspeksiTahunan/{id}/export', [InspeksiP3KController::class, 'exportTahunan_pdf'])->middleware('supervisior');
+    Route::resource('/datap3k', DataP3KController::class)->middleware('supervisior');
+    Route::get('/masterinspeksi/{id}/export', [MasterInspeksiP3KController::class, 'export_pdf'])->middleware('supervisior');
+    Route::resource('/masterinspeksi', MasterInspeksiP3KController::class)->middleware('supervisior');
     Route::get('/inspeksi', [InspeksiP3KController::class, 'index']);
     Route::get('/inspeksi/{id}/inputInpeksi', [InspeksiP3KController::class, 'inputInspeksi']);
     Route::put('/inputInpeksi/{id}', [InspeksiP3KController::class, 'storeInpeksi']);
